pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lakehouse-provider"
version = "0.1.0"
description = "Template for a data provider for the neutron-lakehouse ecosystem"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Your Organization" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyarrow>=14.0.0",
    "duckdb>=0.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
pandas = ["pandas>=2.0.0"]
s3 = ["s3fs>=2024.0.0"]
mcp = ["mcp>=1.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "lakehouse-provider[pandas,s3,mcp]",
]

[project.scripts]
lakehouse-provider-mcp = "lakehouse_provider.mcp:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lakehouse_provider"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/scripts",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
